move pull request modal to a React.FC so it can update state
